+++ features/admin.claims.v1/components/edit/local-claim/edit-basic-details-local-claims.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2023-2025, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,11 +16,18 @@
  * under the License.
  */
 
+import Paper from "@oxygen-ui/react/Paper";
+import Table from "@oxygen-ui/react/Table";
+import TableBody from "@oxygen-ui/react/TableBody";
+import TableCell from "@oxygen-ui/react/TableCell";
+import TableHead from "@oxygen-ui/react/TableHead";
+import TableRow from "@oxygen-ui/react/TableRow";
 import { Show, useRequiredScopes } from "@wso2is/access-control";
 import { AppConstants, AppState, FeatureConfigInterface, history } from "@wso2is/admin.core.v1";
 import useUIConfig from "@wso2is/admin.core.v1/hooks/use-ui-configs";
 import { attributeConfig } from "@wso2is/admin.extensions.v1";
 import { SCIMConfigs } from "@wso2is/admin.extensions.v1/configs/scim";
+import { useGetCurrentOrganizationType } from "@wso2is/admin.organizations.v1/hooks/use-get-organization-type";
 import {
     ConnectorPropertyInterface,
     GovernanceConnectorInterface,
@@ -52,9 +59,11 @@ import {
     DangerZone,
     DangerZoneGroup,
     EmphasizedSegment,
+    Heading,
     Hint,
     Link,
-    Message
+    Message,
+    Tooltip
 } from "@wso2is/react-components";
 import { AxiosError } from "axios";
 import isEmpty from "lodash-es/isEmpty";
@@ -139,6 +148,18 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
     const [ accountVerificationEnabled, setAccountVerificationEnabled ] = useState<boolean>(false);
     const [ selfRegistrationEnabled, setSelfRegistrationEnabledEnabled ] = useState<boolean>(false);
     const [ isSystemClaim, setIsSystemClaim ] = useState<boolean>(false);
+
+    const { isSubOrganization } = useGetCurrentOrganizationType();
+    const [ isConsoleRequired, setIsConsoleRequired ] = useState<boolean>(false);
+    const [ isEndUserRequired, setIsEndUserRequired ] = useState<boolean>(false);
+    const [ isSelfRegistrationRequired, setIsSelfRegistrationRequired ] = useState<boolean>(false);
+    const [ isConsoleReadOnly, setIsConsoleReadOnly ] = useState<boolean>(false);
+    const [ isEndUserReadOnly, setIsEndUserReadOnly ] = useState<boolean>(false);
+    const [ isSelfRegistrationReadOnly, setIsSelfRegistrationReadOnly ] = useState<boolean>(false);
+
+    const isDistinctAttributeProfilesDisabled: boolean = featureConfig?.attributeDialects?.disabledFeatures?.includes(
+        ClaimManagementConstants.DISTINCT_ATTRIBUTE_PROFILES_FEATURE_FLAG
+    );
 
     const { t } = useTranslation();
 
@@ -453,23 +474,78 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
     };
 
     const onSubmit = (values: Record<string, unknown>) => {
-        const data: Claim = {
-            attributeMapping: claim.attributeMapping,
-            claimURI: claim.claimURI,
-            description: values?.description !== undefined ? values.description?.toString() : claim?.description,
-            displayName: values?.name !== undefined ? values.name?.toString() : claim?.displayName,
-            displayOrder: attributeConfig.editAttributes.getDisplayOrder(
-                claim.displayOrder, values.displayOrder?.toString()),
-            properties: claim.properties,
-            readOnly: values?.readOnly !== undefined ? !!values.readOnly : claim?.readOnly,
-            regEx:  values?.regularExpression !== undefined ? values.regularExpression?.toString() : claim?.regEx,
-            required: values?.required !== undefined && !values?.readOnly ? !!values.required : false,
-            sharedProfileValueResolvingMethod: values?.sharedProfileValueResolvingMethod as
-                SharedProfileValueResolvingMethod || SharedProfileValueResolvingMethod.FROM_ORIGIN,
-            supportedByDefault: values?.supportedByDefault !== undefined
-                ? !!values.supportedByDefault : claim?.supportedByDefault,
-            uniquenessScope: values?.uniquenessScope as UniquenessScope || UniquenessScope.NONE
-        };
+        let data: Claim;
+
+        if (isDistinctAttributeProfilesDisabled) {
+            // Use the legacy configuration.
+            data = {
+                attributeMapping: claim.attributeMapping,
+                claimURI: claim.claimURI,
+                description: values?.description !== undefined ? values.description?.toString() : claim?.description,
+                displayName: values?.name !== undefined ? values.name?.toString() : claim?.displayName,
+                displayOrder: attributeConfig.editAttributes.getDisplayOrder(
+                    claim.displayOrder, values.displayOrder?.toString()),
+                properties: claim?.properties,
+                readOnly: values?.readOnly !== undefined ? !!values.readOnly : claim?.readOnly,
+                regEx:  values?.regularExpression !== undefined ? values.regularExpression?.toString() : claim?.regEx,
+                required: values?.required !== undefined && !values?.readOnly ? !!values.required : false,
+                supportedByDefault: values?.supportedByDefault !== undefined
+                    ? !!values.supportedByDefault : claim?.supportedByDefault,
+                uniquenessScope: values?.uniquenessScope as UniquenessScope || UniquenessScope.NONE
+            };
+        } else {
+            // Use the new configuration.
+            data = {
+                attributeMapping: claim.attributeMapping,
+                claimURI: claim.claimURI,
+                description: values?.description !== undefined ? values.description?.toString() : claim?.description,
+                displayName: values?.name !== undefined ? values.name?.toString() : claim?.displayName,
+                displayOrder: attributeConfig.editAttributes.getDisplayOrder(
+                    claim.displayOrder, values.displayOrder?.toString()),
+                profiles: {
+                    console: {
+                        readOnly: values?.consoleReadOnly !== undefined ?
+                            !!values.consoleReadOnly
+                            : claim?.profiles?.console?.readOnly,
+                        required: values?.consoleRequired !== undefined ?
+                            !!values.consoleRequired
+                            : claim?.profiles?.console?.required,
+                        supportedByDefault: values?.consoleSupportedByDefault !== undefined ?
+                            !!values.consoleSupportedByDefault
+                            : claim?.profiles?.console?.supportedByDefault
+                    },
+                    endUser: {
+                        readOnly: values?.endUserReadOnly !== undefined ?
+                            !!values.endUserReadOnly
+                            : claim?.profiles?.endUser?.readOnly,
+                        required: values?.endUserRequired !== undefined ?
+                            !!values.endUserRequired
+                            : claim?.profiles?.endUser?.required,
+                        supportedByDefault: values?.endUserSupportedByDefault !== undefined ?
+                            !!values.endUserSupportedByDefault
+                            : claim?.profiles?.endUser?.supportedByDefault
+                    },
+                    selfRegistration: {
+                        readOnly: values?.selfRegistrationReadOnly !== undefined
+                            ? !!values.selfRegistrationReadOnly
+                            : claim?.profiles?.selfRegistration?.readOnly,
+                        required: values?.selfRegistrationRequired !== undefined ?
+                            !!values.selfRegistrationRequired
+                            : claim?.profiles?.selfRegistration?.required,
+                        supportedByDefault: values?.selfRegistrationSupportedByDefault !== undefined ?
+                            !!values.selfRegistrationSupportedByDefault
+                            : claim?.profiles?.selfRegistration?.supportedByDefault
+                    }
+                },
+                properties: claim?.properties,
+                readOnly: values?.readOnly !== undefined ? !!values.readOnly : claim?.readOnly,
+                regEx:  values?.regularExpression !== undefined ? values.regularExpression?.toString() : claim?.regEx,
+                required: values?.required !== undefined && !values?.readOnly ? !!values.required : false,
+                supportedByDefault: values?.supportedByDefault !== undefined
+                    ? !!values.supportedByDefault : claim?.supportedByDefault,
+                uniquenessScope: values?.uniquenessScope as UniquenessScope || UniquenessScope.NONE
+            };
+        }
 
         if (isSystemClaim || !isUpdatingSharedProfilesEnabled) {
             delete data.sharedProfileValueResolvingMethod;
@@ -505,6 +581,244 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
             .finally(() => {
                 setIsSubmitting(false);
             });
+    };
+
+    const resolveAttributeSupportedByDefaultRow = (): ReactElement => {
+        const isSupportedByDefaultCheckboxDisabled: boolean = !hasMapping
+            || (
+                accountVerificationEnabled
+                && selfRegistrationEnabled
+                && claim?.claimURI ===
+                    ClaimManagementConstants.EMAIL_CLAIM_URI
+                && usernameConfig?.enableValidator === "true"
+            );
+
+        return (
+            <TableRow>
+                <TableCell>
+                    { t("claims:local.forms.profiles.displayByDefault") }
+                    <Tooltip
+                        compact
+                        trigger={ (
+                            <Icon name="info circle" color="grey" className="ml-1" />
+                        ) }
+                        content={
+                            t("claims:local.forms.profiles.displayByDefaultHint")
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Display by default in console"
+                        name="consoleSupportedByDefault"
+                        required={ false }
+                        defaultValue={ claim?.profiles?.console?.supportedByDefault ?? claim?.supportedByDefault }
+                        data-componentid={
+                            `${ testId }-form-console-supported-by-default-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isSupportedByDefaultCheckboxDisabled }
+                        {
+                            ...( isConsoleRequired && !isConsoleReadOnly
+                                ? { checked: true }
+                                : { defaultValue : claim?.profiles?.console?.supportedByDefault
+                                    ?? claim?.supportedByDefault }
+                            )
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Display by default in My Account"
+                        name="endUserSupportedByDefault"
+                        required={ false }
+                        defaultValue={ claim?.profiles?.endUser?.supportedByDefault ?? claim?.supportedByDefault }
+                        data-componentid={ `${ testId }-form-end-user-supported-by-default-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isSupportedByDefaultCheckboxDisabled }
+                        {
+                            ...( isEndUserRequired && !isEndUserReadOnly
+                                ? { checked: true }
+                                : { defaultValue : claim?.profiles?.endUser?.supportedByDefault
+                                    ?? claim?.supportedByDefault }
+                            )
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Display by default in self-registration"
+                        name="selfRegistrationSupportedByDefault"
+                        required={ false }
+                        defaultValue={ claim?.profiles?.selfRegistration?.supportedByDefault ??
+                            claim?.supportedByDefault }
+                        data-componentid={
+                            `${ testId }-form-self-registration-supported-by-default-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isSupportedByDefaultCheckboxDisabled }
+                        {
+                            ...( isSelfRegistrationRequired && !isSelfRegistrationReadOnly
+                                ? { checked: true }
+                                : { defaultValue : claim?.profiles?.selfRegistration?.supportedByDefault ??
+                                    claim?.supportedByDefault }
+                            )
+                        }
+                    />
+                </TableCell>
+            </TableRow>
+        );
+    };
+
+    const resolveAttributeRequiredRow = (): ReactElement => {
+        const isRequiredCheckboxDisabled: boolean = !hasMapping
+            || (
+                accountVerificationEnabled
+                && selfRegistrationEnabled
+                && claim?.claimURI === ClaimManagementConstants.EMAIL_CLAIM_URI
+                && usernameConfig?.enableValidator === "true"
+            );
+
+        return (
+            <TableRow>
+                <TableCell>
+                    { t("claims:local.forms.profiles.required") }
+                    <Tooltip
+                        compact
+                        trigger={ (
+                            <Icon name="info circle" color="grey" className="ml-1" />
+                        ) }
+                        content={
+                            t("claims:local.forms.profiles.requiredHint")
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Required in console"
+                        name="consoleRequired"
+                        required={ false }
+                        defaultValue={ claim?.profiles?.console?.required ?? claim?.required }
+                        data-componentid={ `${ testId }-form-console-required-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isRequiredCheckboxDisabled || isConsoleReadOnly }
+                        listen ={ (value: boolean) => {
+                            setIsConsoleRequired(value);
+                        } }
+                        {
+                            ...( isConsoleReadOnly
+                                ? { value: false }
+                                : { defaultValue : claim?.profiles?.console?.required ?? claim?.required }
+                            )
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Required in My Account"
+                        name="endUserRequired"
+                        required={ false }
+                        defaultValue={ claim?.profiles?.endUser?.required ?? claim?.required }
+                        data-componentid={ `${ testId }-form-end-user-required-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isRequiredCheckboxDisabled || isEndUserReadOnly }
+                        listen ={ (value: boolean) => {
+                            setIsEndUserRequired(value);
+                        } }
+                        {
+                            ...( isEndUserReadOnly
+                                ? { value: false }
+                                : { defaultValue : claim?.profiles?.endUser?.required ?? claim?.required }
+                            )
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Required in self-registration"
+                        name="selfRegistrationRequired"
+                        required={ false }
+                        defaultValue={ claim?.profiles?.selfRegistration?.required ?? claim?.required }
+                        data-componentid={ `${ testId }-form-self-registration-required-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isRequiredCheckboxDisabled || isSelfRegistrationReadOnly }
+                        listen ={ (value: boolean) => {
+                            setIsSelfRegistrationRequired(value);
+                        } }
+                        {
+                            ...( isSelfRegistrationReadOnly
+                                ? { value: false }
+                                : { defaultValue : claim?.profiles?.selfRegistration?.required ?? claim?.required }
+                            )
+                        }
+                    />
+                </TableCell>
+            </TableRow>
+        );
+    };
+
+    const resolveAttributeReadOnlyRow = (): ReactElement => {
+        const isReadOnlyCheckboxDisabled: boolean = !hasMapping;
+
+        return (
+            <TableRow hideBorder>
+                <TableCell>
+                    { t("claims:local.forms.profiles.readonly") }
+                    <Tooltip
+                        compact
+                        trigger={ (
+                            <Icon name="info circle" color="grey" className="ml-1" />
+                        ) }
+                        content={
+                            t("claims:local.forms.profiles.readonlyHint")
+                        }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Read-only in console"
+                        name="consoleReadOnly"
+                        required={ false }
+                        requiredErrorMessage=""
+                        defaultValue={ claim?.profiles?.console?.readOnly ?? claim?.readOnly }
+                        data-componentid={ `${ testId }-form-console-readOnly-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isReadOnlyCheckboxDisabled }
+                        listen ={ (value: boolean) => {
+                            setIsConsoleReadOnly(value);
+                        } }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Read-only in My Account"
+                        name="endUserReadOnly"
+                        required={ false }
+                        requiredErrorMessage=""
+                        defaultValue={ claim?.profiles?.endUser?.readOnly ?? claim?.readOnly }
+                        data-componentid={ `${ testId }-form-end-user-readOnly-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isReadOnlyCheckboxDisabled }
+                        listen ={ (value: boolean) => {
+                            setIsEndUserReadOnly(value);
+                        } }
+                    />
+                </TableCell>
+                <TableCell align="center">
+                    <Field.Checkbox
+                        ariaLabel="Read-only in self-registration"
+                        name="selfRegistrationReadOnly"
+                        required={ false }
+                        requiredErrorMessage=""
+                        defaultValue={ claim?.profiles?.selfRegistration?.readOnly ?? claim?.readOnly }
+                        data-componentid={ `${ testId }-form-self-registration-readOnly-checkbox` }
+                        readOnly={ isSubOrganization() || isReadOnly }
+                        disabled={ isReadOnlyCheckboxDisabled }
+                        listen ={ (value: boolean) => {
+                            setIsSelfRegistrationReadOnly(value);
+                        } }
+                    />
+                </TableCell>
+            </TableRow>
+        );
     };
 
     return (
@@ -555,7 +869,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                         maxLength={ 30 }
                         minLength={ 1 }
                         hint={ t("claims:local.forms.nameHint") }
-                        readOnly={ isReadOnly }
+                        readOnly={ isSubOrganization() || isReadOnly }
 
                     />
                     <Field.Textarea
@@ -574,7 +888,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                         minLength={ 3 }
                         data-testid={ `${ testId }-form-description-input` }
                         hint={ t("claims:local.forms.descriptionHint") }
-                        readOnly={ isReadOnly }
+                        readOnly={ isSubOrganization() || isReadOnly }
                     />
 
                     { !attributeConfig.localAttributes.createWizard.showRegularExpression && !hideSpecialClaims
@@ -593,7 +907,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                 maxLength={ ClaimManagementConstants.REGEX_FIELD_MAX_LENGTH }
                                 minLength={ ClaimManagementConstants.REGEX_FIELD_MIN_LENGTH }
                                 hint={ t("claims:local.forms.regExHint") }
-                                readOnly={ isReadOnly }
+                                readOnly={ isSubOrganization() || isReadOnly }
                             />
                         )
                     }
@@ -623,6 +937,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                 label={ t("claims:local.forms.uniquenessScope.label") }
                                 data-componentid={ `${ testId }-form-uniqueness-scope-dropdown` }
                                 hint={ t("claims:local.forms.uniquenessScopeHint") }
+                                disabled={ isSubOrganization() }
                                 options={ [
                                     {
                                         text: t("claims:local.forms.uniquenessScope.options.none"),
@@ -641,35 +956,31 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                             />
                         )
                     }
-                    { !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI) &&  mappingChecked
+                    { !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI) && mappingChecked
                         ? (
-                            !hideSpecialClaims &&
+                            !hideSpecialClaims && !hasMapping &&
                             (<Grid.Row columns={ 1 } >
                                 <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 14 }>
                                     <Message
                                         type="info"
-                                        content={
-                                            !hasMapping ? (
-                                                <>
-                                                    { t("claims:local.forms.infoMessages." +
-                                                        "disabledConfigInfo") }
-                                                    <div>
-                                                        Add SCIM mapping from
-                                                        <Link
-                                                            external={ false }
-                                                            onClick={ () =>
-                                                                history.push(
-                                                                    AppConstants.getPaths().get("SCIM_MAPPING")
-                                                                )
-                                                            }
-                                                        > here
-                                                        </Link>.
-                                                    </div>
-                                                </>
-                                            ):(
-                                                t("claims:local.forms.infoMessages." +
-                                                    "configApplicabilityInfo")
-                                            )
+                                        content={ (
+                                            <>
+                                                { t("claims:local.forms.infoMessages." +
+                                                    "disabledConfigInfo") }
+                                                <div>
+                                                    Add SCIM mapping from
+                                                    <Link
+                                                        external={ false }
+                                                        onClick={ () =>
+                                                            history.push(
+                                                                AppConstants.getPaths().get("SCIM_MAPPING")
+                                                            )
+                                                        }
+                                                    > here
+                                                    </Link>.
+                                                </div>
+                                            </>
+                                        )
                                         }
                                     />
                                 </Grid.Column>
@@ -679,6 +990,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                     }
                     {
                         //Hides on user_id, username and groups claims
+                        isDistinctAttributeProfilesDisabled &&
                         claim && !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI)
                             && claim.claimURI !== ClaimManagementConstants.USER_ID_CLAIM_URI
                             && claim.claimURI !== ClaimManagementConstants.USER_NAME_CLAIM_URI
@@ -695,7 +1007,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                     setIsShowDisplayOrder(!!values?.supportedByDefault);
                                 } }
                                 data-testid={ `${testId}-form-supported-by-default-input` }
-                                readOnly={ isReadOnly }
+                                readOnly={ isSubOrganization() || isReadOnly }
                                 disabled={
                                     !hasMapping
                                     || (
@@ -737,11 +1049,12 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                 ref={ displayOrderField }
                                 data-testid={ `${ testId }-form-display-order-input` }
                                 hint={ t("claims:local.forms.displayOrderHint") }
-                                readOnly={ isReadOnly }
+                                readOnly={ isSubOrganization() || isReadOnly }
                             />
                         )
                     }
                     {
+                        isDistinctAttributeProfilesDisabled &&
                         claim && !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI)
                             && attributeConfig.editAttributes.showRequiredCheckBox
                             && claim.claimURI !== ClaimManagementConstants.GROUPS_CLAIM_URI
@@ -753,7 +1066,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                 requiredErrorMessage=""
                                 label={ t("claims:local.forms.required.label") }
                                 data-testid={ `${ testId }-form-required-checkbox` }
-                                readOnly={ isReadOnly }
+                                readOnly={ isSubOrganization() || isReadOnly }
                                 hint={ t("claims:local.forms.requiredHint") }
                                 listen ={ (value: boolean) => {
                                     isSupportedByDefault(value);
@@ -792,6 +1105,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                     }
                     {
                         //Hides on user_id, username and groups claims
+                        isDistinctAttributeProfilesDisabled &&
                         claim && !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI)
                             && claim.claimURI !== ClaimManagementConstants.USER_ID_CLAIM_URI
                             && claim.claimURI !== ClaimManagementConstants.USER_NAME_CLAIM_URI
@@ -808,7 +1122,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                     requiredErrorMessage=""
                                     defaultValue={ claim?.readOnly }
                                     data-testid={ `${ testId }-form-readonly-checkbox` }
-                                    readOnly={ isReadOnly }
+                                    readOnly={ isSubOrganization() || isReadOnly }
                                     hint={ t("claims:local.forms.readOnlyHint") }
                                     listen={ (value: boolean) => {
                                         setIsClaimReadOnly(value);
@@ -818,7 +1132,63 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                             )
                     }
                     {
-                        !hideSpecialClaims &&
+                        // Hides on groups claim
+                        !isDistinctAttributeProfilesDisabled && claim && !hideSpecialClaims && mappingChecked
+                            && claim.claimURI !== ClaimManagementConstants.GROUPS_CLAIM_URI && (
+                            <>
+                                <Divider />
+                                <Heading as="h4">
+                                    { t("claims:local.forms.profiles.attributeConfigurations.title") }
+                                </Heading>
+                                <Heading as="h6" color="grey" compact>
+                                    { t("claims:local.forms.profiles.attributeConfigurations.description") }
+                                </Heading>
+                                <Divider hidden />
+                                <Paper variant="outlined" elevation={ 0 }>
+                                    <Table>
+                                        <TableHead>
+                                            <TableRow>
+                                                <TableCell></TableCell>
+                                                <TableCell align="center">
+                                                    { t("claims:local.forms.profiles.administratorConsole") }
+                                                </TableCell>
+                                                <TableCell align="center">
+                                                    { t("claims:local.forms.profiles.endUserProfile") }
+                                                </TableCell>
+                                                <TableCell align="center">
+                                                    { t("claims:local.forms.profiles.selfRegistration") }
+                                                </TableCell>
+                                            </TableRow>
+                                        </TableHead>
+                                        <TableBody>
+                                            {
+                                                // Hides on user_id, username claims
+                                                !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI)
+                                                    && claim.claimURI !== ClaimManagementConstants.USER_ID_CLAIM_URI
+                                                    && claim.claimURI !== ClaimManagementConstants.USER_NAME_CLAIM_URI
+                                                    && resolveAttributeSupportedByDefaultRow()
+                                            }
+                                            {
+                                                !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI)
+                                                    && attributeConfig.editAttributes.showRequiredCheckBox
+                                                    && resolveAttributeRequiredRow()
+                                            }
+                                            {
+                                                // Hides on user_id, username and email claims
+                                                !READONLY_CLAIM_CONFIGS.includes(claim?.claimURI)
+                                                    && claim.claimURI !== ClaimManagementConstants.USER_ID_CLAIM_URI
+                                                    && claim.claimURI !== ClaimManagementConstants.USER_NAME_CLAIM_URI
+                                                    && claim.claimURI !== ClaimManagementConstants.EMAIL_CLAIM_URI
+                                                    && resolveAttributeReadOnlyRow()
+                                            }
+                                        </TableBody>
+                                    </Table>
+                                </Paper>
+                            </>
+                        )
+                    }
+                    {
+                        !hideSpecialClaims && !isSubOrganization() &&
                         (
                             <Show
                                 when={ featureConfig?.attributeDialects?.scopes?.update }