+++ features/admin.groups.v1/components/group-list.tsx
@@ -46,7 +46,7 @@ import React, { ReactElement, ReactNode, SyntheticEvent, useState } from "react"
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Header, Icon, Label, SemanticICONS } from "semantic-ui-react";
-import { GroupConstants } from "../constants";
+import { GroupConstants } from "../constants/group-constants";
 import { GroupsInterface } from "../models/groups";
 
 interface GroupListProps extends SBACInterface<FeatureConfigInterface>,