+++ modules/core/src/models/claim.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020-2024, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2020-2025, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -34,6 +34,20 @@ export interface Claim {
     sharedProfileValueResolvingMethod?: SharedProfileValueResolvingMethod;
     attributeMapping?: AttributeMapping[];
     properties?: Property[];
+    profiles?: {
+        console?: AttributeProfileConfig;
+        endUser?: AttributeProfileConfig;
+        selfRegistration?: AttributeProfileConfig;
+    }
+}
+
+/**
+ * Interface for attribute profile configuration.
+ */
+export interface AttributeProfileConfig {
+    readOnly?: boolean;
+    required?: boolean;
+    supportedByDefault?: boolean;
 }
 
 /**